requirements.txt
fastapi
pandas
geopandas
numpy
uvicorn
python-dotenv
requests
pyproj
typing
geoparquet
pyarrow
boto3
alembic
psycopg2
pydantic
fastapi-sqlalchemy
asyncpg
databases
pyjwt
cryptography
coverage
pytest
black
flake8
pylint
flake8-in-file-ignores
pre-commit